+++ tests/examples/authentication.test.js
@@ -1,0 +1,71 @@
+import 'jest';
+import { map } from 'funcadelic';
+import microstate, * as MS from '../../src';
+
+describe('authentication example', () => {
+  class Session {
+    content = null;
+    constructor(state) {
+      if (state) {
+        return new AuthenticatedSession(state);
+      } else {
+        return new AnonymousSession(state);
+      }
+    }
+  }
+
+  class AuthenticatedSession {
+    isAuthenticated = true;
+    content = MS.Object;
+
+    logout() {
+      return this(AnonymousSession);
+    }
+  }
+
+  class AnonymousSession {
+    content = null;
+    isAuthenticated = false;
+    authenticate(current, user) {
+      return this(AuthenticatedSession, { content: user });
+    }
+  }
+
+  class MyApp {
+    session = Session;
+  }
+
+  describe('AnonymousSession', () => {
+    let ms, authenticated;
+    beforeEach(() => {
+      ms = microstate(MyApp);
+      authenticated = ms.session.authenticate({
+        name: 'Charles',
+      });
+    })
+    it('initializes into AnonymousSession without initial state', () => {
+      expect(ms.state.session).toBeInstanceOf(AnonymousSession);
+    });
+    it('transitions AnonymousSession to Authenticated with authenticate', () => {
+      expect(authenticated.state.session).toBeInstanceOf(AuthenticatedSession);
+      expect(authenticated.state.session).toEqual({
+        content: { name: 'Charles' },
+        isAuthenticated: true,
+      });
+    });
+  });
+
+  describe('AuthenticatedSession', () => {
+    let ms, anonymous;
+    beforeEach(() => {
+      ms = microstate(MyApp, { session: { name: 'Taras' } })
+      anonymous = ms.session.logout();
+    });
+    it('initializes into AuthenticatedSession state', () => {
+      expect(ms.state.session).toBeInstanceOf(AuthenticatedSession);
+    });
+    it('transitions Authenticated session to AnonymousSession with logout', () => {
+      expect(anonymous.state.session).toBeInstanceOf(AnonymousSession);
+    });
+  });
+});
